+++ driver/web/conversions_user.go
@@ -51,7 +51,7 @@ func accountAuthTypesToDef(items []model.AccountAuthType) []Def.AccountAuthTypeF
 //Profile
 func profileFromDef(item *Def.ProfileFields) *model.Profile {
 	return &model.Profile{ID: *item.Id, PhotoURL: *item.PhotoUrl, FirstName: *item.FirstName, LastName: *item.LastName,
-		Email: *item.Email, Phone: *item.Phone, BirthYear: int8(*item.BirthYear), Address: *item.Address, ZipCode: *item.ZipCode,
+		Email: *item.Email, Phone: *item.Phone, BirthYear: int16(*item.BirthYear), Address: *item.Address, ZipCode: *item.ZipCode,
 		State: *item.State, Country: *item.Country}
 }
